# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.1",
        "title": "Property APIs Final",
        "description": "Property Booking API Documentation"
    },
    "host": "localhost:3000",
    "basePath": "/v1",
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/properties": {
            "get": {
                "description": "",
                "summary": "Property with filter",
                "tags": [
                    "Misc"
                ],
                "operationId": "PropertiesGet2",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "bedrooms",
                        "in": "query",
                        "required": true,
                        "type": "integer",
                        "format": "int32",
                        "description": ""
                    },
                    {
                        "name": "purchasePrice",
                        "in": "query",
                        "required": true,
                        "type": "integer",
                        "format": "int32",
                        "description": ""
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "required": true,
                        "type": "boolean",
                        "description": ""
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/PropertyWithFilterBedrooms"
                        },
                        "examples": {
                            "application/json": {
                                "data": [
                                    {
                                        "garden": true,
                                        "_id": "5b49e811fa9f6d0eade076fe",
                                        "title": "Title 2",
                                        "description": "Test description 2",
                                        "typeOfDepartment": "apartment",
                                        "city": "new york",
                                        "numberOfFloors": 5,
                                        "livingSpace": 200,
                                        "totalAreas": 233,
                                        "usableArea": 433,
                                        "rooms": "5",
                                        "bedrooms": 2,
                                        "bathrooms": 4,
                                        "parkingSpace": false,
                                        "purchasePrice": 130000000,
                                        "brokerFee": 20000,
                                        "constructionYear": 2015,
                                        "propertyState": "new",
                                        "heating": true,
                                        "qualityInterior": "Luxurious",
                                        "energyCertificate": true,
                                        "balcony": true,
                                        "availability": true,
                                        "availableDate": "2018-01-02T18:30:00Z",
                                        "availableDuration": 10,
                                        "rentalYield": 5,
                                        "guaranteedRentalYield": true,
                                        "zipCode": 2345,
                                        "direction": "North",
                                        "images": "<link>",
                                        "latitude": 11.33,
                                        "longitude": 13.22,
                                        "address": "Sec 15 New York",
                                        "distanceToCityCenter": "",
                                        "furnished": true,
                                        "storage": true,
                                        "mapping": "iframe",
                                        "studio": true,
                                        "aboutProperty": "",
                                        "propertyManagement": "",
                                        "financing": true,
                                        "welcomeNote": "",
                                        "price": 100021
                                    }
                                ]
                            }
                        }
                    }
                },
                "security": [],
                "x-unitTests": [
                    {
                        "request": {
                            "method": "GET",
                            "uri": "/properties?bedrooms=4&purchasePrice=150000000&filter=true"
                        },
                        "expectedResponse": {
                            "x-allowExtraHeaders": true,
                            "x-bodyMatchMode": "RAW",
                            "x-arrayOrderedMatching": false,
                            "x-arrayCheckCount": false,
                            "x-matchResponseSchema": true,
                            "headers": {
                                "Connection": "keep-alive",
                                "Content-Length": "897",
                                "Content-Type": "application/json; charset=utf-8",
                                "Date": "Sat, 14 Jul 2018 13:31:52 GMT",
                                "ETag": "W/\"381-866DcbBs4OHkpBDvfhiQFtB+JuA\"",
                                "X-Powered-By": "Express"
                            },
                            "body": "{\"data\":[{\"garden\":true,\"_id\":\"5b49e811fa9f6d0eade076fe\",\"title\":\"Title 2\",\"description\":\"Test description 2\",\"typeOfDepartment\":\"apartment\",\"city\":\"new york\",\"numberOfFloors\":5,\"livingSpace\":200,\"totalAreas\":233,\"usableArea\":433,\"rooms\":\"5\",\"bedrooms\":2,\"bathrooms\":4,\"parkingSpace\":false,\"purchasePrice\":130000000,\"brokerFee\":20000,\"constructionYear\":2015,\"propertyState\":\"new\",\"heating\":true,\"qualityInterior\":\"Luxurious\",\"energyCertificate\":true,\"balcony\":true,\"availability\":true,\"availableDate\":\"2018-01-02T18:30:00Z\",\"availableDuration\":10,\"rentalYield\":5,\"guaranteedRentalYield\":true,\"zipCode\":2345,\"direction\":\"North\",\"images\":\"<link>\",\"latitude\":11.33,\"longitude\":13.22,\"address\":\"Sec 15 New York\",\"distanceToCityCenter\":\"\",\"furnished\":true,\"storage\":true,\"mapping\":\"iframe\",\"studio\":true,\"aboutProperty\":\"\",\"propertyManagement\":\"\",\"financing\":true,\"welcomeNote\":\"\",\"price\":100021}]}"
                        },
                        "x-testShouldPass": true,
                        "x-testEnabled": true,
                        "x-testName": "Property with filter",
                        "x-testDescription": ""
                    }
                ],
                "x-operation-settings": {
                    "CollectParameters": false,
                    "AllowDynamicQueryParameters": false,
                    "AllowDynamicFormParameters": false,
                    "IsMultiContentStreaming": false
                }
            }
        },
...
    },
    "definitions": {
...
        "PropertyWithFilterBedrooms": {
            "title": "Property with filter bedrooms",
            "example": {
                "data": [
                    {
                        "garden": true,
                        "_id": "5b49e811fa9f6d0eade076fe",
                        "title": "Title 2",
                        "description": "Test description 2",
                        "typeOfDepartment": "apartment",
                        "city": "new york",
                        "numberOfFloors": 5,
                        "livingSpace": 200,
                        "totalAreas": 233,
                        "usableArea": 433,
                        "rooms": "5",
                        "bedrooms": 2,
                        "bathrooms": 4,
                        "parkingSpace": false,
                        "purchasePrice": 130000000,
                        "brokerFee": 20000,
                        "constructionYear": 2015,
                        "propertyState": "new",
                        "heating": true,
                        "qualityInterior": "Luxurious",
                        "energyCertificate": true,
                        "balcony": true,
                        "availability": true,
                        "availableDate": "2018-01-02T18:30:00Z",
                        "availableDuration": 10,
                        "rentalYield": 5,
                        "guaranteedRentalYield": true,
                        "zipCode": 2345,
                        "direction": "North",
                        "images": "<link>",
                        "latitude": 11.33,
                        "longitude": 13.22,
                        "address": "Sec 15 New York",
                        "distanceToCityCenter": "",
                        "furnished": true,
                        "storage": true,
                        "mapping": "iframe",
                        "studio": true,
                        "aboutProperty": "",
                        "propertyManagement": "",
                        "financing": true,
                        "welcomeNote": "",
                        "price": 100021
                    }
                ]
            },
            "type": "object",
            "properties": {
                "data": {
                    "description": "",
                    "example": [
                        {
                            "garden": true,
                            "_id": "5b49e811fa9f6d0eade076fe",
                            "title": "Title 2",
                            "description": "Test description 2",
                            "typeOfDepartment": "apartment",
                            "city": "new york",
                            "numberOfFloors": 5,
                            "livingSpace": 200,
                            "totalAreas": 233,
                            "usableArea": 433,
                            "rooms": "5",
                            "bedrooms": 2,
                            "bathrooms": 4,
                            "parkingSpace": false,
                            "purchasePrice": 130000000,
                            "brokerFee": 20000,
                            "constructionYear": 2015,
                            "propertyState": "new",
                            "heating": true,
                            "qualityInterior": "Luxurious",
                            "energyCertificate": true,
                            "balcony": true,
                            "availability": true,
                            "availableDate": "2018-01-02T18:30:00Z",
                            "availableDuration": 10,
                            "rentalYield": 5,
                            "guaranteedRentalYield": true,
                            "zipCode": 2345,
                            "direction": "North",
                            "images": "<link>",
                            "latitude": 11.33,
                            "longitude": 13.22,
                            "address": "Sec 15 New York",
                            "distanceToCityCenter": "",
                            "furnished": true,
                            "storage": true,
                            "mapping": "iframe",
                            "studio": true,
                            "aboutProperty": "",
                            "propertyManagement": "",
                            "financing": true,
                            "welcomeNote": "",
                            "price": 100021
                        }
                    ],
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Datum"
                    }
                }
            },
            "required": [
                "data"
            ]
        },
...
    }
}